requirements.txt
asgiref==3.9.1
Django==4.1.4
dj_database_url==0.5.0
psycopg2-binary==2.9.10
whitenoise==6.6.0
gunicorn==20.1.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.1.1
